let bundler decide whether env is node or browser
