+++ src/browser.js
@@ -1,0 +1,4 @@
+const attach = require('./attach');
+const ENVIRONMENT_IS_WORKER = typeof importScripts === 'function';
+
+module.exports = attach(ENVIRONMENT_IS_WORKER ? self : window);
