+++ src/node.js
@@ -1,0 +1,3 @@
+const attach = require('./attach');
+
+module.exports = attach(global);
